First step to separating UI code from program logic
main.c and uemis.c no longer depend on gtk (there are still glib / gobject
dependencies, though).

Eventually the goal is to completely separate the program logic from gtk
so that it becomes easier to port the program to other toolkits or OSs.
