+++ uemis.c
@@ -16,10 +16,7 @@
 #include <time.h>
 #include <regex.h>
 
-#include <gtk/gtk.h>
-
 #include "dive.h"
-#include "display.h"
 #include "uemis.h"
 
 /*
@@ -214,8 +211,11 @@ static void parse_divelog_binary(char *base64, struct dive **divep) {
 	return;
 }
 
+/* parse a single file
+ * TODO: we don't report any errors when the parse fails - we simply don't add them to the list
+ */
 void
-parse_uemis_file(char *divelogfilename,GError **error) {
+parse_uemis_file(char *divelogfilename) {
 	char *found=NULL;
 	struct tm tm;
 	struct dive *dive;
@@ -261,40 +261,7 @@ bail:
  */
 void
 uemis_import() {
-	GtkWidget *dialog;
-	GtkFileFilter *filter = gtk_file_filter_new ();
-	gtk_file_filter_add_pattern (filter, "*.SDA");
-	gtk_file_filter_set_name(filter, "uemis Zurich SDA files");
-	dialog = gtk_file_chooser_dialog_new("Open File",
-					GTK_WINDOW(main_window),
-					GTK_FILE_CHOOSER_ACTION_OPEN,
-					GTK_STOCK_CANCEL, GTK_RESPONSE_CANCEL,
-					GTK_STOCK_OPEN, GTK_RESPONSE_ACCEPT,
-					NULL);
-	gtk_file_chooser_set_select_multiple(GTK_FILE_CHOOSER(dialog), TRUE);
-	gtk_file_chooser_add_filter(GTK_FILE_CHOOSER(dialog),filter);
-
-	if (gtk_dialog_run(GTK_DIALOG(dialog)) == GTK_RESPONSE_ACCEPT) {
-		GSList *filenames;
-		char *filename;
-		filenames = gtk_file_chooser_get_filenames(GTK_FILE_CHOOSER(dialog));
-
-		GError *error = NULL;
-		while(filenames != NULL) {
-			filename = (char *)filenames->data;
-			parse_uemis_file(filename, &error);
-			if (error != NULL)
-			{
-				report_error(error);
-				g_error_free(error);
-				error = NULL;
-			}
-
-			g_free(filename);
-			filenames = g_slist_next(filenames);
-		}
-		g_slist_free(filenames);
+	if (open_import_file_dialog("*.SDA","uemis Zurich SDA files",
+					&parse_uemis_file))
 		report_dives();
-	}
-	gtk_widget_destroy(dialog);
-}
+}
